+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/DecodeOneFrameUtil.java
@@ -35,7 +35,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 
 /** Utilities for decoding a frame for tests. */
 @UnstableApi
-public class DecodeOneFrameTestUtil {
+public class DecodeOneFrameUtil {
 
   /** Listener for decoding events. */
   public interface Listener {
@@ -173,5 +173,5 @@ public class DecodeOneFrameTestUtil {
     }
   }
 
-  private DecodeOneFrameTestUtil() {}
+  private DecodeOneFrameUtil() {}
 }
